requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.1
PyMuPDF>=1.24.0
oracledb>=2.2.0
psycopg[binary]>=3.2.0
